Refactoring of #240
