+++ src/main/java/spark/webserver/MatcherFilter.java
@@ -42,6 +42,7 @@ import spark.route.HttpMethod;
 import spark.route.RouteMatch;
 import spark.route.SimpleRouteMatcher;
 import spark.utils.GzipUtils;
+import spark.webserver.serializing.SerializerChain;
 
 /**
  * Filter for matching of filters and routes.
