+++ src/main/java/spark/webserver/serializing/SerializerChain.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package spark.webserver;
+package spark.webserver.serializing;
 
 import java.io.IOException;
 import java.io.OutputStream;
